Move to float64 to avoid casting for all standard functions.

+++ go/basic_buff.go
@@ -1,18 +1,18 @@
 package codingdojo
 
 type BasicBuff struct {
-	soak   float32
-	damage float32
+	soak   float64
+	damage float64
 }
 
-func makeBasicBuff(soak float32, damage float32) BasicBuff {
+func makeBasicBuff(soak float64, damage float64) BasicBuff {
 	return BasicBuff{soak, damage}
 }
 
-func (this BasicBuff) soakModifier() float32 {
+func (this BasicBuff) soakModifier() float64 {
 	return this.soak
 }
 
-func (this BasicBuff) damageModifier() float32 {
+func (this BasicBuff) damageModifier() float64 {
 	return this.damage
 }
