+++ go/buff.go
@@ -1,6 +1,6 @@
 package codingdojo
 
 type Buff interface {
-	soakModifier() float32
-	damageModifier() float32
+	soakModifier() float64
+	damageModifier() float64
 }
