+++ go/item.go
@@ -3,5 +3,5 @@ package codingdojo
 type Item interface {
 	getBaseDamage() int32
 
-	getDamageModifier() float32
+	getDamageModifier() float64
 }
